Split presentation particles and sector levels
