+++ CMakeLists.txt
@@ -215,7 +215,9 @@ target_sources(
         src/game_presentation_brush_worlds.c
         src/game_presentation_cache_lifecycle.c
         src/game_presentation_menus.c
+        src/game_presentation_particles.c
         src/game_presentation_scene_flow.c
+        src/game_presentation_sector_levels.c
         src/game_data_validation_actions.c
         src/game_data_validation_brush_common.c
         src/game_data_validation_app_world.c
